requirements.txt
requests
tldextract
nltk
python-dotenv
